requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
datasets==3.1.0
dill==0.3.8
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
huggingface-hub==0.30.2
idna==3.10
multidict==6.1.0
multiprocess==0.70.16
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.0
pyarrow==17.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
tokenizers==0.20.3
tqdm==4.67.1
transformers==4.46.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.15.2
PyPDF2==3.0.1
